crypto: Reduce the amount of crypto libraries
This commit removes the `crypto` and `Blake2` module for cryptographic
operations in favor of the `enacl` module.

The `crypto` module is still used, but only in two places in order to
perform an XOR operation on two bitstrings.

Beside this, the runtime for `Crypto.hash/2` got improved significantly,
probably due to the fact that `enacl` calls C code, which generally
happens to be much faster than a higher-level pedant.

Fixes #2